pkg refactor

+++ modelos/pkg/util.py
@@ -10,12 +10,14 @@ def copy_any(src: str, dest: str) -> None:
         src (str): Source filepath
         dest (str): Dest filepath
     """
+    if not os.path.exists(src):
+        raise ValueError(f"src '{src}' does not exist")
     if os.path.isdir(src):
-        shutil.copytree(src, dest)
+        shutil.copytree(src, dest, dirs_exist_ok=True)
     elif os.path.isfile(src):
         shutil.copy(src, dest)
     else:
-        logging.warn(f"skipping path '{src}' as it is not a directory or file")
+        logging.warn(f"@ skipping path '{src}' as it is not a directory or file")
 
     return
 
@@ -26,6 +28,9 @@ def rm_any(filepath: str) -> None:
     Args:
         filepath (str): The filepath to remove
     """
+    if not os.path.exists(filepath):
+        raise ValueError(f"filepath '{filepath}' does not exist")
+
     if os.path.isdir(filepath):
         shutil.rmtree(filepath)
     elif os.path.isfile(filepath):
@@ -34,3 +39,11 @@ def rm_any(filepath: str) -> None:
         logging.warn(f"skipping path '{filepath}' as it is not a directory or file")
 
     return
+
+
+# def tags_to_str(tags: List[str]) -> str:
+#     return ",".join(tags)
+
+
+# def str_to_tags(tags: str) -> List[str]:
+#     return tags.split(",")
